Update and rename 160_Flugans_Stolta_Gång_på_rullgardinen to 160_Flugans_stolta_sång_på_rullgardinen
@StelFoog 

I have  put composer on it's own line, but i would like it to be noted that it is the composer of the melody, not the writer of theese lyrics.

I have fixed the other issues.

Regarding what is more common, my own searches gave me three common results. Flugan, Kräftan and Loppan. Since i have never heard of anyone besides me in IN-sektionen that knows of this, i would say that flugan is the most commonly used in IN. Also, kräftan is more specific for kräftskiva, whie flugan and loppan are more general versions. One could perhaps include the alternate version for Kräftskiva also?

i will gladly accept help with buidinig